requirements.txt
Flask==2.3.2
requests==2.28.1
Pillow==9.2.0
